# github-repo/openapi-schema.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EvokeEssence Exchange API",
    "description": "API schema for EvokeEssence cryptocurrency exchange",
    "version": "1.0.0",
    "contact": {
      "name": "EvokeEssence s.r.o",
      "url": "https://evo-exchange.com",
      "email": "contact@evo-exchange.com"
    }
  },
  "servers": [
    {
      "url": "https://evo-exchange.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/admin/employees": {
      "get": {
        "summary": "Get all employees",
        "description": "Returns a list of all employees (admin only)",
        "operationId": "getAllEmployees",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of employees",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 123
                      },
                      "username": {
                        "type": "string",
                        "example": "employee123"
                      },
                      "email": {
                        "type": "string",
                        "format": "email",
                        "example": "employee@example.com"
                      },
                      "fullName": {
                        "type": "string",
                        "example": "John Smith"
                      },
                      "userGroup": {
                        "type": "string",
                        "example": "finance_emp"
                      },
                      "isActive": {
                        "type": "boolean",
                        "example": true
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-15T08:30:00Z"
                      },
                      "permissions": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        },
                        "example": [
                          "view_clients",
                          "approve_kyc"
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated or not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Admin access required"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create employee",
        "description": "Creates a new employee account (admin only)",
        "operationId": "createEmployee",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "email",
                  "password",
                  "fullName",
                  "userGroup"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "newemployee123"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "newemployee@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "SecurePassword123"
                  },
                  "fullName": {
                    "type": "string",
                    "example": "Jane Smith"
                  },
                  "userGroup": {
                    "type": "string",
                    "enum": [
                      "kyc_employee",
                      "finance_emp",
                      "viewonly_employee",
                      "second_admin"
                    ],
                    "example": "kyc_employee"
                  },
                  "permissions": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "view_clients",
                      "approve_kyc"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Employee created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 124
                    },
                    "username": {
                      "type": "string",
                      "example": "newemployee123"
                    },
                    "message": {
                      "type": "string",
                      "example": "Employee created successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Username already exists"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated or not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Admin access required"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chat/test": {
      "get": {
        "summary": "Test chat service",
        "description": "Tests if the chat service is properly configured",
        "operationId": "testChatService",
        "tags": [
          "Chat"
        ],
        "responses": {
          "200": {
            "description": "Chat service test successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "message": {
                      "type": "string",
                      "example": "The chat service is working correctly"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Chat service error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Chat service is not configured"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Express session cookie for authentication"
      }
    },
...
  },
...
}